db: Add support for reading model data from CSV files
This adds the concept of loading model fixtures from files. Exports might be
given in CSV, YAML, JSON, BSON, or some other format. A Loader class is used
to define the loading strategy. The SimpleLoader simply uses the
ActiveRecord pattern, but one that uses a bulk import strategy could be
written along with database backend support.

A Reader class is used to turn the records from the input file into models
of the target class(es). It is the loader which saves them to the database.