test: Add setup to create and load Northwind
This adds a setup script to create the tables in the database backend and
load the test data. This will make the tests more consistent as the testing
can assume that the data model and initial data exists before any tests are
run. Tests should be careful not to modify the initial data so that other
tests can be built on assumptions based on the data in the CSV files from
which the data is loaded.