db: Fix tests for edges, add SchemaBuilder and ModelManager
This is mostly a checkpoint commit which introduces several key changes:

A ModelManager is associated with Models now which is responsible for
the interface between the Model class and the database backend. It handles
saving and deleting instances, it provides the lookup utility, access to the
queryset, and can be used to tie instances to a particular backend (which
would be nice for testing and migration).

A SchemaBuilder is used now allow the Model to define its own fields,
declaratively. This is not required, but is helpful for the model to define
strongly typed fields. This was previously inspected from the database (and
is still supported); however using the SchemaBuilder is likely faster, more
direct, and less magical. It is also helpful for installations, where the
fields do not need to be specified elsewhere to create the tables, and will
help with automatic migration creation in the future.

Some backends such as SQLite do not provide a mechanism to determine
the exact field types of tables, which make type coercion for import
problematic (i.e., 1 != '1': it's required for WHERE clauses to work
correctly. The data on both sides of an expression match type-wise). If
used, the SchemaBuilder allows models to connect the fields of their tables
with specific types.