transform: Separate into Lookup and Transform
A Transform is an operation to change a query filter expression into SQL.
This might include virtual columns like `year` for the DateTime field or
something more creating like the JSON field. A Lookup is an extension to
the Transform class which is specifically used to compare values. Using the
two allows for `exact` to be silently implied to the end of Transform
instances in filter expressions.

Also, finally get a working concept for the JSON field. And fix the failing
tests.