json: Add support for objects/arrays
This is added as a property for the field, so specifying `object=true` in the
setup properties for the field will result in json_decode creating a stdClass
object. This is more difficult to deal with in PHP and requires odd parentheses
and so is disabled by default. Instead, decoded json is presented and
interacted with as an array.