joins: More strides toward support of composite FKs
This patch eliminates use of the local_pk and foreign_pk fields and tries
not to assume that the foreign key is associated with exactly one field in
both models. However, a comparison operation still needs to be defined in
order to build a lookup with a composite key. It would need to build a tuple
and send such to the database instead of a simple value.