fields: ForeignKey can automatically add joins
A model does not need to specify both `joins` in the static meta data and
ForeignKey fields. Instead, it can specify the join name and the reverse
join name (`reverse_name`) in the options for a ForeignKey field. When
inspected, the ForeignKey will add the joins to the local and foreign models
automatically.

ForeignKey now extends from BaseField, but delegates the database
interactions to the referenced field type. This will allow for proper type
conversion of data types for queries involving ForeignKey fields without
repeating the foreign key column. That is, the ForeignKey field will emulate
the field which it references.