joins: Add the ability to bypass the through table
For many-to-many relationships (edges), the middle table is bypassed when
used in path expressions. This is a good thing in that it matches closer to
the way most people approach m2m relationships; however, in order to filter
based on fields in the middle model, a second join is required -- either
explicitly or from the `reverse_name` of the ForeignKey in the middle model.