queryset: Fix circular reference error
This fixes an error where the ModelInstanceManager maintained a reference to
the QuerySet instance, and the QuerySet instance managed a reference to the
ModelInstanceManager instance (if it's the iterator for the query). Because
of the circular reference, if the iterator is not exhausted, then the
resource is not closed and the query remains open. This wastes memory and
prevents some other queries from running after such a situation happens.

QuerySetTest::testOffset generates such a case. It sets up a query to fetch
ten results but only fetches one. In this case, the query remains open and
both the QuerySet and ModelInstanceManager instances are not cleaned up.
Therefore, the associated driver and statement do not release the database
cursor and resources. Later, if something like a DDL statement is executed
on the table, the database may refuse to alter a table for which there is an
open cursor.

At the conclusion of the test suite, the database is torn down. On SQLite at
least, the DROP TABLE statements are refused because the tables remain
locked for the open query.

This addresses the issue by removing the circular reference between the
QuerySet and the ModelInstanceManager.