test: Fix corruption of loaded models
The previous technique would create a single instance and clone it for
each model created or loaded from database. This introduces a different
technique where the ReflectionClass instance is cached and a new
instance is retrieved from it for each model loaded from the database.
This fixes the possible case where the reference instance would somehow
get corrupted and then cloned.